move lomap_old_mapper to conftest
